+++ spec/generators/encrypted_columns_generator_spec.rb
@@ -1,0 +1,49 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require 'generators/doorkeeper/encrypted_columns_generator'
+
+describe 'Doorkeeper::EncryptedColumnsGenerator' do
+  include GeneratorSpec::TestCase
+
+  tests Doorkeeper::EncryptedColumnsGenerator
+  destination ::File.expand_path('../tmp/dummy', __FILE__)
+
+  describe 'after running the generator' do
+    before :each do
+      prepare_destination
+    end
+
+    context 'pre Rails 5.0.0' do
+      it 'creates a migration with no version specifier' do
+        stub_const("ActiveRecord::VERSION::MAJOR", 4)
+        stub_const("ActiveRecord::VERSION::MINOR", 2)
+
+        run_generator
+
+        assert_migration 'db/migrate/add_encrypted_columns.rb' do |migration|
+          assert migration.include?("ActiveRecord::Migration\n")
+          assert migration.include?(':encrypted_secret')
+          assert migration.include?(':encrypted_token')
+          assert migration.include?(':encrypted_refresh_token')
+        end
+      end
+    end
+
+    context 'post Rails 5.0.0' do
+      it 'creates a migration with a version specifier' do
+        stub_const("ActiveRecord::VERSION::MAJOR", 5)
+        stub_const("ActiveRecord::VERSION::MINOR", 0)
+
+        run_generator
+
+        assert_migration 'db/migrate/add_encrypted_columns.rb' do |migration|
+          assert migration.include?("ActiveRecord::Migration[5.0]\n")
+          assert migration.include?(':encrypted_secret')
+          assert migration.include?(':encrypted_token')
+          assert migration.include?(':encrypted_refresh_token')
+        end
+      end
+    end
+  end
+end