+++ spec/support/shared/hashing_shared_context.rb
@@ -27,3 +27,13 @@ shared_context 'with application hashing enabled' do
     end
   end
 end
+
+shared_context 'with encryption enabled' do
+  let(:hashed_or_plain_token_func) { Doorkeeper::Application.method(:hashed_or_plain_token) }
+  before do
+    Doorkeeper.configure do
+      encrypt_token_secrets
+      encryption_key '0c829d3227f04bf8c751bab43fa35ca8925b4615bf18472856e1a004c1081269'
+    end
+  end
+end